+++ src/main/resources/com/asesinosdesoftware/javeparking/MenuOperarioView.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<AnchorPane fx:id="contenedor" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.asesinosdesoftware.javeparking.controller.MenuOperarioViewController">
+<AnchorPane fx:id="contenedor" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.asesinosdesoftware.javeparking.controller.menu.MenuOperarioViewController">
     <children>
         <Button layoutX="247.0" layoutY="218.0" mnemonicParsing="false" prefHeight="29.0" prefWidth="103.0" text="Pago" onAction="#IngresoPagoOp">
             <font>
