+++ src/main/resources/com/asesinosdesoftware/javeparking/PagoOpView.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/17.0.2-ea" fx:controller="com.asesinosdesoftware.javeparking.controller.PagoOpController">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/17.0.2-ea" fx:controller="com.asesinosdesoftware.javeparking.controller.operario.PagoOpController">
     <children>
         <TextField fx:id="placaTextField" layoutX="200.0" layoutY="50.0" prefWidth="200.0" promptText="Ingrese la placa" />
         <ComboBox fx:id="metodoPagoComboBox" layoutX="200.0" layoutY="100.0" prefWidth="200.0">
